Stencil time
(30/I/15)
This was an interesting experiment. After drawing some sketches, and
tracing them into illustrator, I hit the joyful dead end of letterform
tweaking. I felt that it was easy to lose sight of the big picture
here, working into the teardrop terminal of the a, and the gap of the
c.

Overall, the letters feel clunky, and whatever spirit was used in
making the originals was lost.

This sounds disparaging, however, it was a good trial and error
experiment. One notable thing was the use of the pencil tool in
illustrator to tweak shapes. In tandem with the brush, there could be a
lot of letterform possibilities to be played with, so we’ll see what
happens.